Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=0.27.0
crewai
langgraph
pdfplumber
python-docx
pillow
pytesseract
pydub
typing_extensions
